Timer and Game Over
Implemented a basic Timer and a Game Over screen based on which player won. If one of the players' health reaches 0, the other player wins. Alternatively if the timer reaches 0 the player with the most health wins. In case of a tie, the match result is pronounced as a tie.